+++ fili-core/src/main/java/com/yahoo/bard/webservice/druid/model/query/LookbackQuery.java
@@ -266,20 +266,7 @@ public class LookbackQuery extends AbstractDruidAggregationQuery<LookbackQuery> 
 
     @Override
     public LookbackQuery withIntervals(Collection<Interval> intervals) {
-        return new LookbackQuery(
-                new QueryDataSource(getInnerQueryUnchecked().withIntervals(intervals)),
-                granularity,
-                filter,
-                aggregations,
-                postAggregations,
-                intervals,
-                context,
-                true,
-                lookbackOffsets,
-                lookbackPrefixes,
-                having,
-                limitSpec
-        );
+        return withDataSource(new QueryDataSource(getInnerQueryUnchecked().withIntervals(intervals)));
     }
 
     @Override
@@ -322,6 +309,10 @@ public class LookbackQuery extends AbstractDruidAggregationQuery<LookbackQuery> 
 
     public LookbackQuery withLookbackOffsets(List<Period> lookbackOffsets) {
         return new LookbackQuery(dataSource, granularity, filter, aggregations, postAggregations, intervals, context, false, lookbackOffsets, lookbackPrefixes, having, limitSpec);
+    }
+
+    public LookbackQuery withDoFork(boolean doFork) {
+        return new LookbackQuery(dataSource, granularity, filter, aggregations, postAggregations, intervals, context, doFork, lookbackOffsets, lookbackPrefixes, having, limitSpec);
     }
     // CHECKSTYLE:ON